requirements.txt
# Core LangChain
langchain
langchain-core
langchain-community



# Environment Variables
python-dotenv


# Google (Gemini)
google-generativeai
langchain-google-genai

# Machine Learning Utilities
numpy
scikit-learn

